+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Support to specify lcid for Export-PnPTaxonomy for a particular termset
 - Added support in the Navigation cmdlets to manage the site footer on modern sites.
 - Added Invoke-PnPSPRestMethod cmdlet to execute REST requests towards a SharePoint site.
+- Added Enable-PnPCommSite cmdlet to convert the root site collection of a tenant into a communication site.
 
 ### Changed